fix(ci): split subagent API owners
