+++ anyharness/crates/anyharness-lib/src/api/mod.rs
@@ -3,12 +3,17 @@ pub mod http;
 pub mod openapi;
 pub mod router;
 pub mod sse;
+mod subagents_openapi;
 pub mod ws;
 
 #[cfg(test)]
 mod openapi_tests;
 #[cfg(test)]
+mod review_admission_tests;
+#[cfg(test)]
 mod router_tests;
+#[cfg(test)]
+mod session_admission_tests;
 #[cfg(test)]
 mod workflow_runs_portable_contract_tests;
 #[cfg(test)]
@@ -17,7 +22,3 @@ mod workflow_runs_scripted_tests;
 mod workflow_runs_tests;
 #[cfg(test)]
 mod workflow_workspaces_tests;
-#[cfg(test)]
-mod session_admission_tests;
-#[cfg(test)]
-mod review_admission_tests;
