+++ anyharness/crates/anyharness-lib/src/api/openapi.rs
@@ -3,9 +3,6 @@ use anyharness_contract::v1::{
     AgentCredentialState, AgentInstallProgress, AgentInstallProgressComponent,
     AgentInstallProgressPhase, AgentInstallState, AgentLaunchModelOption, AgentLaunchOption,
     AgentLaunchOptionsResponse, AgentLoginTerminalRecord, AgentLoginTerminalStatus,
-    AgentOperationsAgent, AgentOperationsCapability, AgentOperationsConfiguration,
-    AgentOperationsExecutionStatus, AgentOperationsIdentity, AgentOperationsPresentationStatus,
-    AgentOperationsRole, AgentOperationsStatus, AgentOperationsWorkspaceIdentity,
     AgentReadinessState, AgentSeedFailureKind, AgentSeedHealth, AgentSeedLastAction,
     AgentSeedOwnership, AgentSeedSource, AgentSeedStatus, AgentSummary,
     ApplyAgentAuthStateResponse, ArtifactStatus, AvailableCommandsUpdatePayload,
@@ -86,14 +83,12 @@ use anyharness_contract::v1::{
     SessionMcpEnvVar, SessionMcpHeader, SessionMcpHttpServer, SessionMcpServer,
     SessionMcpStdioServer, SessionMcpTransport, SessionRawNotificationEnvelope,
     SessionReviewsResponse, SessionStartedEvent, SessionStateUpdatePayload, SessionStatus,
-    SessionSubagentsResponse, SetSessionConfigOptionRequest, SetSessionConfigOptionResponse,
-    SetSessionGoalRequest, SetSessionLoopRequest, SetupHint, SetupHintCategory,
-    SetupScriptExecution, SetupScriptStatus, StagePatchRequest, StagePathsRequest,
-    StartAgentLoginRequest, StartAgentLoginResponse, StartAgentLoginTerminalResponse,
-    StartCodeReviewRequest, StartPlanReviewRequest, StartTerminalCommandRequest,
-    StartTerminalCommandResponse, StartWorkspaceSetupRequest, StatWorkspaceFileResponse,
-    StopReason, SubagentLatestCompletion, SubagentLifecycleResponse, SubagentParentRoster,
-    SubagentRelationship, SubagentRosterEntry, SubagentTurnCompletedPayload, SubagentTurnOutcome,
+    SetSessionConfigOptionRequest, SetSessionConfigOptionResponse, SetSessionGoalRequest,
+    SetSessionLoopRequest, SetupHint, SetupHintCategory, SetupScriptExecution, SetupScriptStatus,
+    StagePatchRequest, StagePathsRequest, StartAgentLoginRequest, StartAgentLoginResponse,
+    StartAgentLoginTerminalResponse, StartCodeReviewRequest, StartPlanReviewRequest,
+    StartTerminalCommandRequest, StartTerminalCommandResponse, StartWorkspaceSetupRequest,
+    StatWorkspaceFileResponse, StopReason, SubagentTurnCompletedPayload, SubagentTurnOutcome,
     TerminalCommandOutputMode, TerminalCommandRunDetail, TerminalCommandRunStatus,
     TerminalCommandRunSummary, TerminalLifecycleEvent, TerminalPurpose, TerminalRecord,
     TerminalStatus, TranscriptItemDeltaPayload, TranscriptItemKind, TranscriptItemPayload,
@@ -114,11 +109,11 @@ use anyharness_contract::v1::{
     WorkspaceMobilitySessionCandidate, WorkspacePurgeOutcome, WorkspacePurgePreflightResponse,
     WorkspacePurgeResponse, WorkspaceRetireBlocker, WorkspaceRetireBlockerCode,
     WorkspaceRetireBlockerSeverity, WorkspaceRetireOutcome, WorkspaceRetirePreflightResponse,
-    WorkspaceRetireResponse, WorkspaceSubagentsResponse, WorkspaceSurface, WorktreeBaseFetch,
-    WorktreeCheckoutMode, WorktreeGitStatusState, WorktreeGitStatusSummary,
-    WorktreeInventoryAction, WorktreeInventoryResponse, WorktreeInventoryRow,
-    WorktreeInventoryState, WorktreeInventoryWorkspaceSummary, WorktreeNameConflictPolicy,
-    WorktreeStorageEstimate, WriteWorkspaceFileRequest, WriteWorkspaceFileResponse,
+    WorkspaceRetireResponse, WorkspaceSurface, WorktreeBaseFetch, WorktreeCheckoutMode,
+    WorktreeGitStatusState, WorktreeGitStatusSummary, WorktreeInventoryAction,
+    WorktreeInventoryResponse, WorktreeInventoryRow, WorktreeInventoryState,
+    WorktreeInventoryWorkspaceSummary, WorktreeNameConflictPolicy, WorktreeStorageEstimate,
+    WriteWorkspaceFileRequest, WriteWorkspaceFileResponse,
 };
 use utoipa::OpenApi;
 
@@ -242,11 +237,6 @@ use utoipa::OpenApi;
         super::http::sessions_lifecycle::restore_dismissed_session,
         super::http::sessions::list_sessions,
         super::http::sessions::get_session,
-        super::http::subagents::get_session_subagents,
-        super::http::subagents::get_workspace_subagents,
-        super::http::subagents::close_subagent,
-        super::http::subagents::open_subagent,
-        super::http::subagents::promote_subagent,
         super::http::sessions_events::list_session_events,
         super::http::sessions_events::list_session_raw_notifications,
         super::http::sessions_interactions::resolve_interaction,
@@ -425,22 +415,6 @@ use utoipa::OpenApi;
         SessionMcpBindingOutcome,
         SessionMcpBindingNotAppliedReason,
         SessionMcpBindingSummary,
-        SessionSubagentsResponse,
-        WorkspaceSubagentsResponse,
-        SubagentLifecycleResponse,
-        SubagentParentRoster,
-        SubagentRosterEntry,
-        SubagentRelationship,
-        SubagentLatestCompletion,
-        AgentOperationsAgent,
-        AgentOperationsIdentity,
-        AgentOperationsWorkspaceIdentity,
-        AgentOperationsRole,
-        AgentOperationsConfiguration,
-        AgentOperationsStatus,
-        AgentOperationsPresentationStatus,
-        AgentOperationsExecutionStatus,
-        AgentOperationsCapability,
         CreateSessionRequest,
         ResumeSessionRequest,
         UpdateSessionTitleRequest,
@@ -706,6 +680,7 @@ pub struct ApiDoc;
 
 pub fn openapi_json() -> String {
     ApiDoc::openapi()
+        .merge_from(super::subagents_openapi::SubagentApiDoc::openapi())
         .to_pretty_json()
         .expect("OpenAPI serialization should not fail")
 }
