+++ anyharness/crates/anyharness-lib/src/app/mod.rs
@@ -558,10 +558,9 @@ impl AppState {
         review_runtime
             .clone()
             .spawn_background_tasks(review_hook_event_rx);
-        let runtime_identity = runtime_identity(&auth_manager, &runtime_home);
         let agent_operations =
             agent_operations::wire_agent_operations(agent_operations::AgentOperationsWiringDeps {
-                runtime_identity,
+                runtime_identity: runtime_identity(&auth_manager, &runtime_home),
                 session_service: session_service.clone(),
                 session_link_service: Arc::new(session_link_service.clone()),
                 subagent_service: subagent_service.clone(),
