+++ anyharness/crates/anyharness-lib/src/domains/agent_operations/mod.rs
@@ -7,3 +7,4 @@
 pub mod mcp;
 pub mod model;
 pub mod runtime;
+pub mod subagents;
