+++ scripts/max_lines_allowlist.txt
@@ -8,8 +8,8 @@
 # when the file is under its threshold.
 
 anyharness/crates/anyharness-contract/src/v1/events.rs 1092 existing AnyHarness contract oversized file
-anyharness/crates/anyharness-contract/src/v1/sessions.rs 779 existing AnyHarness contract oversized file (+caller-selected idempotent create id)
-anyharness/crates/anyharness-lib/src/api/openapi.rs 692 existing AnyHarness oversized file (+hosting schema registration, +agent-auth state response, +gateway catalog endpoints) + enriched gateway-models schema fields (model-table §1) + workflow-runs + local materialization operations + workflow-run-workspaces paths/schemas + agent install progress schemas + worktree restore path/schemas + WorktreeBaseFetch + RuntimeDiskPressure schemas (central OpenAPI registry; per-route path/schema registration only)
+anyharness/crates/anyharness-contract/src/v1/sessions.rs 696 existing AnyHarness contract oversized file (+caller-selected idempotent create id; stale fingerprint reduced after legacy public subagent response removal)
+anyharness/crates/anyharness-lib/src/api/openapi.rs 686 existing AnyHarness oversized file (+hosting schema registration, +agent-auth state response, +gateway catalog endpoints) + enriched gateway-models schema fields (model-table §1) + workflow-runs + local materialization operations + workflow-run-workspaces paths/schemas + agent install progress schemas + worktree restore path/schemas + WorktreeBaseFetch + RuntimeDiskPressure schemas (subagent registrations extracted to their owned OpenAPI document)
 anyharness/crates/anyharness-lib/src/api/router_tests.rs 1898 existing AnyHarness oversized test file (+hosting PR-status route tests, +catalog version tests, +managed default-placement route test, +model-snapshot status route shape/auth tests, +agent-auth/install poke contract tests, +agent-auth/install poke-suppression proofs against the automatic_poke_engine seam (F-036 cycle-2 review item 1); the catalog-push handler tests are gone with the heartbeat transport)
 anyharness/crates/anyharness-lib/src/domains/workflows/service.rs 664 workflow durable rules (+run-control cancel-intent use case and helpers); split into service/ on next growth per guides/domains.md
 anyharness/crates/anyharness-lib/src/domains/workflows/service_tests.rs 1320 workflow-runs merge-gated domain suite (+2b partial-uniqueness contract on nonterminal session binding)
@@ -32,7 +32,7 @@ anyharness/crates/anyharness-lib/src/domains/workspaces/retire_preflight.rs 613 
 anyharness/crates/anyharness-lib/src/live/sessions/probe.rs 679 catalog probe with per-harness model-source fallbacks (+the three runtime-probe options: env removal, scratch workspace root, config-option-only switch suppression); shrank from 683 after the readiness resolve moved to the call boundary (grid plan PR 5c/7 — live never fetches)
 anyharness/crates/anyharness-lib/src/persistence/custom_migrations.rs 1126 custom migration with legacy schema coverage (+0063 partial unique controller index, ordered after the 0062 rebuild)
 anyharness/sdk/src/client/core.ts 603 AnyHarness TS client core (+preserved FastAPI detail body on AnyHarnessError for gateway billing-block classification, C8)
-anyharness/sdk-react/src/hooks/sessions.ts 723 existing SDK React oversized file
+anyharness/sdk-react/src/hooks/sessions.ts 675 existing SDK React oversized file (subagent queries and lifecycle mutations extracted to their owned hook module)
 cloud/sdk/src/types/generated.ts 795 existing generated cloud SDK type surface (+workspaceKind, git numstat fields, OpenAPI-derived materialization types, app-extra overlay, and PR 5 exact-ref create request fields; -actionBlock fields; + lostAt (C7))
 anyharness/sdk/src/reducer/__tests__/transcript.test.ts 1817 existing SDK oversized test file
 anyharness/sdk/src/reducer/transcript.ts 1528 existing SDK oversized file
@@ -71,7 +71,7 @@ apps/packages/product-client/src/domain/chats/transcript/transcript-presentation
 apps/packages/product-client/src/domain/workspaces/cloud-work-inventory.test.ts 859 relocated existing oversized domain test; no growth
 apps/packages/product-client/src/domain/sessions/activity.test.ts 660 relocated existing oversized domain test; no growth
 apps/packages/product-client/src/domain/sessions/intents/session-intent.test.ts 606 relocated existing oversized domain test; no growth
-anyharness/crates/anyharness-lib/src/api/router.rs 640 AnyHarness router (+goals/loops/activity/feed routes, +gateway catalog version route, +workflow-runs/cancel + local materialization operations + workflow-run-workspaces routes + worktree restore route)
+anyharness/crates/anyharness-lib/src/api/router.rs 632 AnyHarness router (+goals/loops/activity/feed routes, +gateway catalog version route, +workflow-runs/cancel + local materialization operations + workflow-run-workspaces routes + worktree restore route; subagent routes composed from their HTTP owner)
 anyharness/crates/anyharness-lib/src/app/mod.rs 669 AnyHarness app wiring (+gateway model planner wiring; the gateway model resolver + probe-engine store wiring is gone with the old-chain deletion, A9) (+session mutation admission purge/retention wiring, +goal/loop/activity/feed services, +two-phase workflow-run and workflow-workspace wiring, +worktree restore runtime wiring; local materialization service wiring is now the materialization runtime valve, moved to app/materialization.rs per grid PR 7; -3: WorkspaceService construction/pass-through deleted, grid PR 8; net 0 in grid PR 9: WorkspaceRetireService joins AppState while the preflight-checker/purge/retire constructions move to the app/workspaces.rs wiring family); PR #1753 extracts process/runtime configuration helpers to app/config.rs while adding Workspace MCP composition; PR #3 moves Agent Operations graph construction to app/agent_operations.rs
 anyharness/crates/anyharness-lib/src/domains/agents/catalog/service_tests.rs 605 B-2 validation-cut rewrote the refusal assertions in place (single SESSION_MODEL_UNSUPPORTED); shrinks when the legacy gateway-models routes delete (C-track cutover)
 anyharness/crates/anyharness-lib/src/domains/loops/runtime.rs 802 loops runtime — native write path + emulated Codex scheduler (+2b derived-safe classification comment at the fire seam; fire executor moved to its owning module in PR5)
